Refactor llm paths
