+++ README.md
@@ -119,7 +119,7 @@ Eval-Framework provides a unified interface for evaluating language models acros
    ```python
     from pathlib import Path
 
-    from eval_framework.llm.huggingface_llm import HFLLM
+    from eval_framework.llm.huggingface import HFLLM
     from eval_framework.main import main
     from eval_framework.tasks.eval_config import EvalConfig
     from template_formatting.formatter import HFFormatter
