+++ docs/evaluate_huggingface_model.md
@@ -8,7 +8,7 @@ Here's a sample of code to evaluate a HuggingFace model:
 
 ```python
 from pathlib import Path
-from eval_framework.llm.huggingface_llm import HFLLM
+from eval_framework.llm.huggingface import HFLLM
 from eval_framework.main import main
 from eval_framework.tasks.eval_config import EvalConfig
 from template_formatting.formatter import HFFormatter
