+++ examples/local_evaluation/huggingface_model_evaluation.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from eval_framework.llm.huggingface_llm import HFLLM
+from eval_framework.llm.huggingface import HFLLM
 from eval_framework.main import main
 from eval_framework.tasks.eval_config import EvalConfig
 from template_formatting.formatter import HFFormatter
